Work around OS X 10.8 issue breaking speech voice changes.
In OS X 10.5 and 10.6, you could change the voice without recreating the world; now you can't (at least, not in the general case).

A similar issue is described here:

http://lists.apple.com/archives/speech-dev/2013/Jan/msg00002.html

To direct speech to an audio device, this replaces AUGraph-based speech synthesis with a speech channel for which an AudioDeviceID is set.  (This API was added in 10.6, whereas the speech AU was added in 10.5, which explains why I couldn't do it this way originally.)