Preliminary fix for voice selection problems.
Any voice which had a “Compact” version did not get its VoiceSpec created, therefore the default voice was being used instead. Now do the matching by ID and creator, although this is not entirely documented to work.

Filter the voice list to match the one in System Preferences. I cannot find a documented way to do this but the list is completely unmanageable otherwise.