Do better at selecting a default voice.
The existence (or not) of the speaker was a bad way of testing whether we’re being set, so just set the default voice at startup all the time.

At some point we might want to persist the selected voice in NSUserDefaults, but for the moment, this is good enough.