Rewrite alias selection in NJRMediaPopUpButton.
Fixes issue in which the recent alias data and menu items get out of sync when an alias needs to be removed.

Now regenerates the alias data on removal, but it’s more intention-revealing and hopefully unlikely to be a problem in practice.

Rewrite item moving to make it clearer what’s going on.

Display a better error message when aliases fail to resolve.

Display a document-modal dialog in case validation happens while the Set Alarm panel is invisible.

Migrate off deprecated NSBeginAlertSheet to NSAlert based API.