Switch Read Me to Auto Layout and remove background parsing.
Pester 1.1b22 had regressed quite substantially and was no longer enforcing minimum or maximum split view widths.

Replaces a lot of finicky code with constraints which generally make sense!

Macs are fast enough now and it turned out that I was doing a lot of stuff in AppKit from a background thread that I shouldn’t be.

Also note that making the content view layer-backed magically fixed the lack of a rounded bottom right corner.