Update build instructions.
This is one of the few times recently when the latest version of Pester does not build cleanly on the latest macOS, so explain how to get to a working build.

Unless I need to do a quick bug fix, 1.1b23 will be the last version of Pester to support OS X 10.9.